SHared:Components:Navigation: move from common and rewrite to typescript
