--- packages/common/components/Navigation/index.js
@@ -1,1 +1,0 @@
-export default from "./Navigation";
