--- packages/common/components/Navigation/sub-components/arrow-btn.js
@@ -1,27 +1,0 @@
-import React from "react";
-import { IconButton } from "@docspace/shared/components/icon-button";
-
-import ArrowPathReactSvgUrl from "PUBLIC_DIR/images/arrow.path.react.svg?url";
-
-const ArrowButton = ({ isRootFolder, onBackToParentFolder }) => {
-  return (
-    <>
-      {!isRootFolder ? (
-        <div className="navigation-arrow-container">
-          <IconButton
-            iconName={ArrowPathReactSvgUrl}
-            size="17"
-            isFill={true}
-            onClick={onBackToParentFolder}
-            className="arrow-button"
-          />
-          <div className="navigation-header-separator" />
-        </div>
-      ) : (
-        <></>
-      )}
-    </>
-  );
-};
-
-export default React.memo(ArrowButton);
