+++ packages/shared/components/context-menu/index.tsx
@@ -2,8 +2,16 @@ import {
   ContextMenuModel,
   ContextMenuType,
   SeparatorType,
+  TGetContextMenuModel,
+  TContextMenuRef,
 } from "./ContextMenu.types";
 
-export type { ContextMenuModel, ContextMenuType, SeparatorType };
+export type {
+  TContextMenuRef,
+  TGetContextMenuModel,
+  ContextMenuModel,
+  ContextMenuType,
+  SeparatorType,
+};
 
 export { ContextMenu } from "./ContextMenu";
