+++ packages/shared/components/navigation/Navigation.tsx
@@ -1,22 +1,21 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-import StyledContainer from "./StyledNavigation";
-import ArrowButton from "./sub-components/arrow-btn";
-import Text from "./sub-components/text";
-import ControlButtons from "./sub-components/control-btn";
-import DropBox from "./sub-components/drop-box";
-
-import { Consumer, DomHelpers } from "@docspace/shared/utils";
-
-import { Backdrop } from "@docspace/shared/components/backdrop";
-
+import React, { useCallback } from "react";
 import { ReactSVG } from "react-svg";
 
-import ToggleInfoPanelButton from "./sub-components/toggle-infopanel-btn";
-import TrashWarning from "./sub-components/trash-warning";
-import NavigationLogo from "./sub-components/logo-block";
-import { DeviceType } from "@docspace/shared/enums";
+import { Consumer, DomHelpers } from "../../utils";
+import { DeviceType } from "../../enums";
+
+import { Backdrop } from "../backdrop";
+
+import ArrowButton from "./sub-components/ArrowBtn";
+import Text from "./sub-components/Text";
+import ControlButtons from "./sub-components/ControlBtn";
+import ToggleInfoPanelButton from "./sub-components/ToggleInfoPanelBtn";
+import TrashWarning from "./sub-components/TrashWarning";
+import NavigationLogo from "./sub-components/LogoBlock";
+import DropBox from "./sub-components/DropBox";
+
+import { StyledContainer } from "./Navigation.styled";
+import { INavigationProps } from "./Navigation.types";
 
 const Navigation = ({
   tReady,
@@ -25,7 +24,7 @@ const Navigation = ({
   title,
   canCreate,
   isTabletView,
-  personal,
+
   onClickFolder,
   navigationItems,
   getContextOptionsPlus,
@@ -55,61 +54,64 @@ const Navigation = ({
   rootRoomTitle,
 
   ...rest
-}) => {
+}: INavigationProps) => {
   const [isOpen, setIsOpen] = React.useState(false);
   const [firstClick, setFirstClick] = React.useState(true);
   const [dropBoxWidth, setDropBoxWidth] = React.useState(0);
-  const [maxHeight, setMaxHeight] = React.useState(false);
+  // const [maxHeight, setMaxHeight] = React.useState("");
 
-  const dropBoxRef = React.useRef(null);
-  const containerRef = React.useRef(null);
+  const dropBoxRef = React.useRef<HTMLDivElement | null>(null);
+  const containerRef = React.useRef<HTMLDivElement | null>(null);
 
   const isDesktop = currentDeviceType === DeviceType.desktop;
 
   const infoPanelIsVisible = React.useMemo(
     () => isDesktop && (!isEmptyPage || (isEmptyPage && isRoom)),
-    [isDesktop, isEmptyPage, isRoom]
+    [isDesktop, isEmptyPage, isRoom],
   );
 
-  const onMissClick = React.useCallback(
-    (e) => {
-      e.preventDefault;
-      const path = e.path || (e.composedPath && e.composedPath());
-
-      if (!firstClick) {
-        !path.includes(dropBoxRef.current) ? toggleDropBox() : null;
-      } else {
-        setFirstClick((prev) => !prev);
-      }
-    },
-    [firstClick, toggleDropBox, setFirstClick]
-  );
-
-  const onClickAvailable = React.useCallback(
-    (id, isRootRoom) => {
-      onClickFolder && onClickFolder(id, isRootRoom);
-      toggleDropBox();
-    },
-    [onClickFolder, toggleDropBox]
-  );
-
-  const toggleDropBox = () => {
+  const toggleDropBox = useCallback(() => {
     if (navigationItems.length === 0) return;
     if (isRootFolder) return setIsOpen(false);
     setIsOpen((prev) => !prev);
 
-    setDropBoxWidth(DomHelpers.getOuterWidth(containerRef.current));
+    if (containerRef.current)
+      setDropBoxWidth(DomHelpers.getOuterWidth(containerRef.current));
 
-    const { top } = DomHelpers.getOffset(containerRef.current);
+    // const { top } = DomHelpers.getOffset(containerRef.current);
 
-    setMaxHeight(`calc(100vh - ${top}px)`);
+    // setMaxHeight(`calc(100vh - ${top}px)`);
 
     setFirstClick(true);
-  };
+  }, [isRootFolder, navigationItems.length]);
+
+  const onMissClick = React.useCallback(
+    (e: MouseEvent) => {
+      e.preventDefault();
+      const path = e.composedPath && e.composedPath();
+
+      if (!firstClick) {
+        if (dropBoxRef.current && !path.includes(dropBoxRef.current))
+          toggleDropBox();
+      } else {
+        setFirstClick((prev) => !prev);
+      }
+    },
+    [firstClick, toggleDropBox, setFirstClick],
+  );
+
+  const onClickAvailable = React.useCallback(
+    (id: number | string, isRootRoom: boolean) => {
+      onClickFolder?.(id, isRootRoom);
+      toggleDropBox();
+    },
+    [onClickFolder, toggleDropBox],
+  );
 
   const onResize = React.useCallback(() => {
-    setDropBoxWidth(DomHelpers.getOuterWidth(containerRef.current));
-  }, [containerRef.current]);
+    if (containerRef.current)
+      setDropBoxWidth(DomHelpers.getOuterWidth(containerRef.current));
+  }, []);
 
   React.useEffect(() => {
     if (isOpen) {
@@ -129,7 +131,7 @@ const Navigation = ({
 
   const onBackToParentFolderAction = React.useCallback(() => {
     setIsOpen((val) => !val);
-    onBackToParentFolder && onBackToParentFolder();
+    onBackToParentFolder?.();
   }, [onBackToParentFolder]);
 
   const showRootFolderNavigation =
@@ -146,9 +148,15 @@ const Navigation = ({
         isOpen={isOpen}
         isRootFolder={isRootFolder}
         onClick={toggleDropBox}
+        isRootFolderTitle={showRootFolderTitle}
       />
     </div>
   );
+
+  const onTextClick = React.useCallback(() => {
+    onClickFolder(navigationItems[navigationItems.length - 2].id, false);
+    setIsOpen(false);
+  }, [navigationItems, onClickFolder]);
 
   const navigationTitleContainerNode = showRootFolderNavigation ? (
     <div className="title-container">
@@ -160,15 +168,7 @@ const Navigation = ({
         isOpen={isOpen}
         isRootFolder={isRootFolder}
         isRootFolderTitle
-        onClick={() => {
-          {
-            onClickFolder(
-              navigationItems[navigationItems.length - 2].id,
-              false
-            );
-            setIsOpen(false);
-          }
-        }}
+        onClick={onTextClick}
       />
       {navigationTitleNode}
     </div>
@@ -185,7 +185,7 @@ const Navigation = ({
               <Backdrop
                 visible={isOpen}
                 withBackground={false}
-                withoutBlur={true}
+                withoutBlur
                 zIndex={400}
               />
 
@@ -193,14 +193,10 @@ const Navigation = ({
                 {...rest}
                 isDesktop={isDesktop}
                 ref={dropBoxRef}
-                maxHeight={maxHeight}
                 dropBoxWidth={dropBoxWidth}
-                sectionHeight={context.sectionHeight}
-                showText={showText}
+                sectionHeight={context.sectionHeight || 0}
                 isRootFolder={isRootFolder}
                 onBackToParentFolder={onBackToParentFolderAction}
-                title={title}
-                personal={personal}
                 canCreate={canCreate}
                 navigationItems={navigationItems}
                 getContextOptionsFolder={getContextOptionsFolder}
@@ -210,10 +206,8 @@ const Navigation = ({
                 isInfoPanelVisible={isInfoPanelVisible}
                 onClickAvailable={onClickAvailable}
                 isDesktopClient={isDesktopClient}
-                showRootFolderNavigation={showRootFolderNavigation}
                 withLogo={withLogo}
                 burgerLogo={burgerLogo}
-                titleIcon={titleIcon}
                 currentDeviceType={currentDeviceType}
                 navigationTitleContainerNode={navigationTitleContainerNode}
               />
@@ -221,10 +215,8 @@ const Navigation = ({
           )}
           <StyledContainer
             ref={containerRef}
-            width={context.sectionWidth}
+            width={context.sectionWidth || 0}
             isRootFolder={isRootFolder}
-            canCreate={canCreate}
-            isTabletView={isTabletView}
             isTrashFolder={isTrashFolder}
             isDesktop={isDesktop}
             isDesktopClient={isDesktopClient}
@@ -235,7 +227,6 @@ const Navigation = ({
             {withLogo && (
               <NavigationLogo
                 className="navigation-logo"
-                logo={withLogo}
                 burgerLogo={burgerLogo}
               />
             )}
@@ -247,13 +238,11 @@ const Navigation = ({
             {navigationTitleContainerNode}
 
             <ControlButtons
-              personal={personal}
               isRootFolder={isRootFolder}
               canCreate={canCreate}
               getContextOptionsFolder={getContextOptionsFolder}
               getContextOptionsPlus={getContextOptionsPlus}
               isEmptyFilesList={isEmptyFilesList}
-              clearTrash={clearTrash}
               toggleInfoPanel={toggleInfoPanel}
               isInfoPanelVisible={isInfoPanelVisible}
               isDesktop={isDesktop}
@@ -266,10 +255,7 @@ const Navigation = ({
             />
           </StyledContainer>
           {isDesktop && isTrashFolder && !isEmptyPage && (
-            <TrashWarning
-              title={titles.trashWarning}
-              isTabletView={isTabletView}
-            />
+            <TrashWarning title={titles.trashWarning} />
           )}
           {infoPanelIsVisible && !hideInfoPanel && (
             <ToggleInfoPanelButton
@@ -284,24 +270,6 @@ const Navigation = ({
       )}
     </Consumer>
   );
-};
-
-Navigation.propTypes = {
-  tReady: PropTypes.bool,
-  isRootFolder: PropTypes.bool,
-  title: PropTypes.string,
-  canCreate: PropTypes.bool,
-  isDesktop: PropTypes.bool,
-  isTabletView: PropTypes.bool,
-  personal: PropTypes.bool,
-  onClickFolder: PropTypes.func,
-  navigationItems: PropTypes.arrayOf(PropTypes.object),
-  getContextOptionsPlus: PropTypes.func,
-  getContextOptionsFolder: PropTypes.func,
-  onBackToParentFolder: PropTypes.func,
-  titles: PropTypes.object,
-  isEmptyPage: PropTypes.bool,
-  isRoom: PropTypes.bool,
 };
 
 export default React.memo(Navigation);
