+++ packages/shared/components/navigation/sub-components/ControlBtn.tsx
@@ -1,0 +1,131 @@
+import React from "react";
+
+import { StyledControlButtonContainer } from "../Navigation.styled";
+import { IControlButtonProps } from "../Navigation.types";
+
+import ToggleInfoPanelButton from "./ToggleInfoPanelBtn";
+import PlusButton from "./PlusBtn";
+import ContextButton from "./ContextBtn";
+
+const ControlButtons = ({
+  isRootFolder,
+  canCreate,
+  getContextOptionsFolder,
+  getContextOptionsPlus,
+  isEmptyFilesList,
+  isInfoPanelVisible,
+  toggleInfoPanel,
+  toggleDropBox,
+  isDesktop,
+  titles,
+  withMenu,
+  onPlusClick,
+  isFrame,
+  isPublicRoom,
+  isTrashFolder,
+  isMobile,
+}: IControlButtonProps) => {
+  const toggleInfoPanelAction = () => {
+    toggleInfoPanel?.();
+    toggleDropBox?.();
+  };
+
+  return (
+    <StyledControlButtonContainer isFrame={isFrame}>
+      {!isRootFolder || (isTrashFolder && !isEmptyFilesList) ? (
+        <>
+          {!isMobile && canCreate && (
+            <PlusButton
+              className="add-button"
+              getData={getContextOptionsPlus}
+              withMenu={withMenu}
+              onPlusClick={onPlusClick}
+              isFrame={isFrame}
+              title={titles?.actions}
+            />
+          )}
+
+          {/* <ContextMenuButton
+            id="header_optional-button"
+            zIndex={402}
+            className="option-button"
+            directionX="right"
+            iconName={VerticalDotsReactSvgUrl}
+            size={15}
+            isFill
+            getData={getContextOptionsFolder}
+            isDisabled={false}
+            title={titles?.contextMenu}
+          /> */}
+
+          <ContextButton
+            id="header_optional-button"
+            className="option-button"
+            getData={getContextOptionsFolder}
+            withMenu={withMenu}
+            // onPlusClick={onPlusClick}
+            title={titles?.actions}
+            isTrashFolder={isTrashFolder}
+            isMobile={isMobile || false}
+          />
+
+          {!isDesktop && (
+            <ToggleInfoPanelButton
+              isRootFolder={isRootFolder}
+              isInfoPanelVisible={isInfoPanelVisible}
+              toggleInfoPanel={toggleInfoPanelAction}
+              titles={titles}
+            />
+          )}
+        </>
+      ) : canCreate ? (
+        <>
+          {!isMobile && (
+            <PlusButton
+              id="header_add-button"
+              className="add-button"
+              getData={getContextOptionsPlus}
+              withMenu={withMenu}
+              onPlusClick={onPlusClick}
+              isFrame={isFrame}
+              title={titles?.actions}
+            />
+          )}
+          {!isDesktop && (
+            <ToggleInfoPanelButton
+              isRootFolder={isRootFolder}
+              isInfoPanelVisible={isInfoPanelVisible}
+              toggleInfoPanel={toggleInfoPanelAction}
+              titles={titles}
+            />
+          )}
+        </>
+      ) : (
+        <>
+          {!isDesktop && (
+            <ToggleInfoPanelButton
+              isRootFolder={isRootFolder}
+              isInfoPanelVisible={isInfoPanelVisible}
+              toggleInfoPanel={toggleInfoPanelAction}
+              titles={titles}
+            />
+          )}
+
+          {isPublicRoom && (
+            <ContextButton
+              id="header_optional-button"
+              className="option-button"
+              getData={getContextOptionsFolder}
+              withMenu={withMenu}
+              title={titles?.contextMenu}
+              isTrashFolder={isTrashFolder}
+              isMobile={isMobile || false}
+            />
+          )}
+        </>
+      )}
+    </StyledControlButtonContainer>
+  );
+};
+
+export default React.memo(ControlButtons);
