+++ packages/shared/components/navigation/sub-components/TrashWarning.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { StyledTrashWarning } from "../Navigation.styled";
+
+const TrashWarning = ({ title }: { title?: string }) => {
+  return (
+    <StyledTrashWarning className="trash-warning">{title}</StyledTrashWarning>
+  );
+};
+
+export default TrashWarning;
